requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
annotated-types==0.6.0
Appium-Python-Client==3.1.0
attrs==23.1.0
backports.functools-lru-cache==1.6.6
blinker==1.7.0
browserstack-local==1.2.2
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
cleo==2.0.1
click==8.1.7
config==0.5.1
crashtest==0.4.1
cryptography==41.0.3
curlify==2.2.1
distlib==0.3.7
dulwich==0.21.6
exceptiongroup==1.1.2
execnet==2.0.2
filelock==3.12.4
Flask==3.0.0
future==0.18.3
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.17.3
jsonschema-specifications==2023.7.1
keyring==24.2.0
MarkupSafe==2.1.3
more-itertools==10.1.0
msgpack==1.0.7
outcome==1.2.0
packaging==23.1
pexpect==4.8.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.2.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
psutil==5.9.6
ptyprocess==0.7.0
py==1.11.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyOpenSSL==23.2.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.0
pytest-xdist==3.3.1
python-dotenv==1.0.0
rapidfuzz==2.15.2
referencing==0.30.2
requests==2.31.0
requests-toolbelt==1.0.0
rpds-py==0.10.6
selene==2.0.0rc4
selenium==4.14.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
trove-classifiers==2023.10.18
typing_extensions==4.8.0
urllib3==2.0.6
urllib3-secure-extra==0.1.0
virtualenv==20.24.6
webdriver-manager==3.8.6
Werkzeug==3.0.1
wsproto==1.2.0
xattr==0.10.1
zipp==3.17.0
